requirements.txt
torch
torchvision
matplotlib
scikit-learn
kaggle
numpy
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
pydantic>=2.0.0